requirements.txt
msgpack-python
celery[redis]
#matplotlib
flask
deep-translator
python-docx 